updated ui logic

+++ app/src/main/java/com/techmave/fisherville/model/Temp.kt
@@ -5,11 +5,11 @@ import com.google.gson.annotations.SerializedName
    
 data class Temp (
 
-   @SerializedName("day") var day : Double,
-   @SerializedName("min") var min : Double,
-   @SerializedName("max") var max : Double,
-   @SerializedName("night") var night : Double,
-   @SerializedName("eve") var eve : Double,
-   @SerializedName("morn") var morn : Double
+   @SerializedName("day") var day : Float,
+   @SerializedName("min") var min : Float,
+   @SerializedName("max") var max : Float,
+   @SerializedName("night") var night : Float,
+   @SerializedName("eve") var eve : Float,
+   @SerializedName("morn") var morn : Float
 
 )
